tests/wpt/tests/css/cssom/CSSRuleList.html

<!doctype html>
<html>
<head>
    <meta charset="utf-8">
    <title>CSSOM - CSSRuleList interface</title>
    <link rel="help" href="https://drafts.csswg.org/cssom/#the-cssrulelist-interface">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <style>
        body { width: 50%; }
        #foo { height: 100px; }
    </style>

    <script>
    test(function () {
        var ruleList = document.styleSheets[0].cssRules;
        assert_equals(ruleList.length, 2, "CSSRuleList length attribute");
        assert_equals(ruleList[0].cssText, "body { width: 50%; }", "CSSRuleList indexed getter");
        assert_equals(ruleList[1].cssText, "#foo { height: 100px; }", "CSSRuleList indexed getter");
        assert_equals(ruleList[2], undefined, "CSSRuleList indexed getter");
        assert_equals(ruleList.item(0).cssText, "body { width: 50%; }", "CSSRuleList item function");
        assert_equals(ruleList.item(1).cssText, "#foo { height: 100px; }", "CSSRuleList item function");
        assert_equals(ruleList.item(2), null, "CSSRuleList item function");
    });
    </script>
</head>
</html>
